+++ TokanPages.ClientApp/src/Shared/Components/index.ts
@@ -5,8 +5,8 @@ export { RenderContent } from "./RenderContent/renderContent";
 export { RenderCardMedia } from "./CustomCardMedia/customCardMedia";
 export { RenderImage } from "./CustomImage/customImage";
 export { GetIcon } from "./GetIcon/getIcon";
-export { RenderSideMenu } from "./MenuRender/renderSideMenu";
-export { RenderNavbarMenu } from "./MenuRender/renderNavbarMenu";
+export { RenderSideMenu } from "./RenderMenu/renderSideMenu";
+export { RenderNavbarMenu } from "./RenderMenu/renderNavbarMenu";
 export { ProgressBarView as ProgressBar } from "./ProgressBar";
 export { ScrollToTop, HideOnScroll, ProgressOnScroll, ClearPageStart } from "./Scroll";
 export { ApplicationToast } from "./Toasts/applicationToast";
